+++ amfiteatr_core/src/env/generic/hashmap_environment_t.rs
@@ -84,7 +84,7 @@ StatefulEnvironment<DP> for TracingHashMapEnvironment<DP, S,C>{
             Ok(updates) => {
 
                 //self.history.push_trace_step(EnvironmentTraceStep::new(state_clone, agent.clone(), action.clone(), true));
-                self.history.register_step(state_clone, agent.clone(), action.clone(), true)?;
+                self.history.register_step_point(state_clone, agent.clone(), action.clone(), true)?;
                 if self.base_environment.state().is_finished(){
                     //self.history.finalize(self.base_environment.state().clone());
                     self.history.finish(self.base_environment.state().clone())?;
@@ -98,7 +98,7 @@ StatefulEnvironment<DP> for TracingHashMapEnvironment<DP, S,C>{
                     //self.history.finalize(self.base_environment.state().clone());
                     self.history.finish(self.base_environment.state().clone())?;
                 }
-                self.history.register_step(state_clone.clone(), agent.clone(), action.clone(), false)?;
+                self.history.register_step_point(state_clone.clone(), agent.clone(), action.clone(), false)?;
                 Err(e)
             }
         }
@@ -119,7 +119,7 @@ ScoreEnvironment<DP> for TracingHashMapEnvironment<DP, S, C>{
             Ok(updates) => {
 
 
-                self.history.register_step(state_clone, agent.clone(), action.clone(), true)?;
+                self.history.register_step_point(state_clone, agent.clone(), action.clone(), true)?;
                 if self.base_environment.state().is_finished(){
                     self.history.finish(self.base_environment.state().clone())?
                 }
@@ -127,7 +127,7 @@ ScoreEnvironment<DP> for TracingHashMapEnvironment<DP, S, C>{
             }
             Err(e) => {
 
-                self.history.register_step(state_clone, agent.clone(), action.clone(), false)?;
+                self.history.register_step_point(state_clone, agent.clone(), action.clone(), false)?;
                 if self.base_environment.state().is_finished(){
                     //self.history.finalize(self.base_environment.state().clone());
                     self.history.finish(self.base_environment.state().clone())?