+++ amfiteatr_core/src/env/trajectory.rs
@@ -1,161 +1,17 @@
 use std::fmt::{Display, Formatter};
-use std::ops::Index;
+use crate::agent::{AgentTrajectory, EvaluatedInformationSet};
 use crate::env::EnvironmentStateSequential;
 use crate::domain::DomainParameters;
 use crate::error::{AmfiteatrError, TrajectoryError};
 
-/*
-/// Trace step of while game (traced by game environment)
-#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
-#[derive(Debug, Clone)]
-pub struct EnvironmentTraceStep<DP: DomainParameters, S: EnvironmentStateSequential<DP>>{
-    state_before: S,
-    agent: DP::AgentId,
-    action: DP::ActionType,
-    is_action_valid: bool
-}
-
-impl<DP: DomainParameters, S: EnvironmentStateSequential<DP>> Display for EnvironmentTraceStep<DP, S>
-where S: Display, <DP as DomainParameters>::AgentId: Display,
-      <DP as DomainParameters>::ActionType: Display{
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        write!(f, "[ {} ][ {} / {} ", self.state_before, self.agent, self.action)?;
-        if !self.is_action_valid{
-            write!(f, "!]")
-        } else {
-            write!(f, "]")
-        }
-    }
-}
 
 
-impl<DP: DomainParameters, S: EnvironmentStateSequential<DP>> EnvironmentTraceStep<DP, S>{
-
-    pub fn new(state_before: S, agent: DP::AgentId,
-               action: DP::ActionType, is_valid: bool) -> Self{
-        /*let checked_action = match is_valid{
-            false => CheckedAction::Invalid(action),
-            true => CheckedAction::Valid(action)
-        };
-
-         */
-        Self{state_before, agent, action, is_action_valid: is_valid}
-    }
-
-    pub fn state_before(&self) -> &S{
-        &self.state_before
-    }
-
-    pub fn agent(&self) -> &DP::AgentId{
-        &self.agent
-    }
-
-    pub fn action(&self) -> &DP::ActionType{
-        &self.action
-    }
-
-
-    pub fn is_action_valid(&self) -> bool{
-        self.is_action_valid
-    }
-}
-
-/// Standard trajectory for environment
-//pub type StdEnvironmentTrajectory<DP, S> = Trajectory<EnvironmentTraceStep<DP, S>>;
-
-#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
-#[derive(Debug)]
-pub struct EnvironmentTrajectory<DP: DomainParameters, S: EnvironmentStateSequential<DP>> {
-
-
-    #[cfg_attr(feature = "serde",
-        serde(bound(serialize =
-            "DP::ActionType: serde::Serialize, \
-            DP::UniversalReward:serde::Serialize, \
-            DP::AgentId: serde::Serialize"
-        )))
-    ]
-    #[cfg_attr(feature = "serde",
-        serde(bound(deserialize =
-            "DP::ActionType: serde::Deserialize<'de>, \
-            DP::UniversalReward: serde::Deserialize<'de>, \
-            DP::AgentId: serde::Deserialize<'de>"
-        )))
-    ]
-    history: Vec<EnvironmentTraceStep<DP, S>>,
-    //revoked_steps: Vec<AgentTraceStep<DP, S>>,
-    final_state: Option<S>,
-
-}
-
-impl<DP: DomainParameters, S: EnvironmentStateSequential<DP>> Default for EnvironmentTrajectory<DP, S>{
-    fn default() -> Self {
-        Self{
-            history: Vec::new(),
-            final_state: None
-        }
-    }
-}
-impl<DP: DomainParameters, S: EnvironmentStateSequential<DP>>  EnvironmentTrajectory<DP, S>
-{
-
-
-    pub fn new() -> Self{
-        Self{ history: Default::default(), final_state: Default::default()}
-    }
-    /*pub fn register_line(&mut self, state: S, action: DP::ActionType, reward_for_action: S::RewardType){
-        self.trace.push(GameTraceLine::new(state, action, reward_for_action));
-
-    }*/
-    pub fn new_reserve(capacity: usize) -> Self{
-        Self{ history: Vec::with_capacity(capacity), final_state: Default::default()}
-    }
-
-    /// Pushes trace step on the end of trajectory.
-    pub fn push_trace_step(&mut self, trace_step: EnvironmentTraceStep<DP, S>){
-        self.history.push(trace_step);
-    }
-    /// Clears trajectory using [`Vec::clear()`](std::vec::Vec::clear)
-    pub fn clear(&mut self){
-        self.history.clear();
-        self.final_state = None;
-    }
-
-    /// Returns reference to `Vec` inside the structure.
-    pub fn list(&self) -> &Vec<EnvironmentTraceStep<DP, S>>{
-        &self.history
-    }
-
-    /// Pops step from trajectory using [`Vec::pop()`](std::vec::Vec::pop)
-    pub fn pop_step(&mut self) -> Option<EnvironmentTraceStep<DP, S>>{
-        self.history.pop()
-    }
-
-
-    pub fn is_empty(&self) -> bool{
-        self.list().is_empty()
-    }
-
-    pub fn finalize(&mut self, state: S){
-        self.final_state = Some(state);
-    }
-
-    pub fn final_state(&self) -> &Option<S>{
-        &self.final_state
-    }
-
-}
-
-impl<DP: DomainParameters, S: EnvironmentStateSequential<DP>> Index<usize> for EnvironmentTrajectory<DP, S>{
-    type Output = EnvironmentTraceStep<DP, S>;
-
-    fn index(&self, index: usize) -> &Self::Output {
-        &self.history[index]
-    }
-}
-
- */
-
+/// View of single step in game trajectory.
+///
+/// Provides references to states
+/// both for time moment before step and at the moment before next step.
+/// View contains also action done and agent performing it in this step.
+/// Game step is measured from one state to the next (transaction is made by any agent performing action).
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 #[derive(Clone, Debug)]
 pub struct GameStepView<'a, DP: DomainParameters, S: EnvironmentStateSequential<DP>>{
@@ -198,14 +54,21 @@ impl<'a, DP: DomainParameters, S: EnvironmentStateSequential<DP>> GameStepView<'
         Self{state_before, agent, action, is_action_valid: is_valid, state_after}
     }
 
-    pub fn state_before(&self) -> &S{
+    /// Returns state before taking action.
+    pub fn state(&self) -> &S{
         &self.state_before
     }
+    /// Returns state after taking action.
+    pub fn late_state(&self) -> &S{
+        &self.state_after
+    }
 
+    /// Returns reference to the agent id who performed action in the step's early state.
     pub fn agent(&self) -> &DP::AgentId{
         &self.agent
     }
 
+    /// Returns reference to action performed during the step
     pub fn action(&self) -> &DP::ActionType{
         &self.action
     }
@@ -215,6 +78,31 @@ impl<'a, DP: DomainParameters, S: EnvironmentStateSequential<DP>> GameStepView<'
         self.is_action_valid
     }
 }
+
+
+/// Game trajectory, keeping information about following _states_, and _actions_ performed by _agents_.
+///
+/// Trajectory consists of series game step point tuples of:
+/// + __state__ in the time just before __agent__ performing __action__,
+/// + __agent__ that performed action,
+/// + __action__ being performed,
+/// + boolean evaluation of __action__ being legal (it usually should be legal, however in case it isn't
+/// it may be noted here.
+///
+/// Trajectory ends with final state.
+/// Data can be considered with following array:
+/// | State | Agent | Action | valid |
+/// |:-----:|:-----:|:------:|:-------|
+/// | s\_0 | A | a\_0 | true |
+/// | s\_1 | B | a\_1 | true |
+/// | ... | ... | ... | true |
+/// | s\_n | B| a\_n | true |
+/// | final state | --- | --- | --- |
+///
+/// /// Each row is added via function [`GameTrajectory::register_step_point`] and the final row is added
+/// via [`GameTrajectory::finish`].
+///
+/// For trace collected by players refer to [`AgentTrajectory`].
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 #[derive(Clone, Debug)]
 pub struct GameTrajectory<DP: DomainParameters, S: EnvironmentStateSequential<DP>>{
@@ -245,10 +133,22 @@ impl<DP: DomainParameters, S: EnvironmentStateSequential<DP>> Default for GameTr
 }
 impl<DP: DomainParameters, S: EnvironmentStateSequential<DP>> GameTrajectory<DP, S>{
 
+    /// Creates new trajectory. Initializing undergoing vectors with [`default`](Default::default).
+    /// If size can be estimated before it will be better to use [`Self::with_capacity`]
     pub fn new() -> Self{
         Default::default()
     }
-
+    /// Creates new trajectory with size hint to allocate memory at initialization.
+    pub fn with_capacity(size: usize) -> Self{
+        Self{
+            states: Vec::with_capacity(size),
+            agents: Vec::with_capacity(size),
+            actions: Vec::with_capacity(size),
+            validations: Vec::with_capacity(size),
+            final_state: None
+        }
+    }
+    /// Clears vectors of data. Uses [`Vec::clear`], so elements are removed but no reallocation occurs.
     pub fn clear(&mut self){
         self.actions.clear();
         self.states.clear();
@@ -256,14 +156,58 @@ impl<DP: DomainParameters, S: EnvironmentStateSequential<DP>> GameTrajectory<DP,
         self.final_state = None;
     }
 
-    pub fn completed_len(&self) -> usize{
+    /// Number of full steps (available for view). Step is complete when there is next information set
+    /// and payoff able to select. This means that initially trajectory has step len 0, after
+    /// first registration step len is still 0, as only initial information set is known and
+    /// not information set at the end of step. Subsequent updates increment number of steps and finishing also.
+    ///
+    /// # Example:
+    /// ```
+    /// use amfiteatr_core::agent::AgentTrajectory;
+    /// use amfiteatr_core::comm::StdEnvironmentEndpoint;
+    /// use amfiteatr_core::demo::{DemoAction, DemoDomain, DemoInfoSet, DemoState};
+    /// use amfiteatr_core::env::GameTrajectory;
+    /// let state = DemoState::new_with_players(vec![2.0, 3.0], 1, &Default::default());
+    /// let mut trajectory: GameTrajectory<DemoDomain, DemoState> = GameTrajectory::new();
+    /// assert_eq!(trajectory.number_of_steps(), 0);
+    /// trajectory.register_step_point(state.clone(), 0, DemoAction(0), true).unwrap();
+    /// assert_eq!(trajectory.number_of_steps(), 0);
+    /// trajectory.register_step_point(state.clone(), 0, DemoAction(1), true).unwrap();
+    /// assert_eq!(trajectory.number_of_steps(), 1);
+    /// trajectory.finish(state.clone()).unwrap();
+    /// assert_eq!(trajectory.number_of_steps(), 2);
+    ///
+    /// ```
+    pub fn number_of_steps(&self) -> usize{
         match self.is_finished(){
             true => self.states.len(),
             false => self.states.len().saturating_sub(1)
         }
     }
-    pub fn register_step(&mut self, state: S, agent: DP::AgentId, action: DP::ActionType, is_valid: bool)
-                         -> Result<(), AmfiteatrError<DP>>{
+    /// Returns the number of raw action points registered. For finished trajectory it equals number
+    /// of steps in game. For unfinished games, the number of steps is one less, as the noted step point
+    /// has no follower.
+    /// # Example:
+    /// ```
+    /// use amfiteatr_core::agent::AgentTrajectory;
+    /// use amfiteatr_core::comm::StdEnvironmentEndpoint;
+    /// use amfiteatr_core::demo::{DemoAction, DemoDomain, DemoInfoSet, DemoState};
+    /// use amfiteatr_core::env::GameTrajectory;
+    /// let state = DemoState::new_with_players(vec![2.0, 3.0], 1, &Default::default());
+    /// let mut trajectory: GameTrajectory<DemoDomain, DemoState> = GameTrajectory::new();
+    /// assert_eq!(trajectory.number_of_action_points(), 0);
+    /// trajectory.register_step_point(state.clone(), 0, DemoAction(0), true).unwrap();
+    /// assert_eq!(trajectory.number_of_action_points(), 1);
+    /// trajectory.register_step_point(state.clone(), 0, DemoAction(1), true).unwrap();
+    /// assert_eq!(trajectory.number_of_action_points(), 2);
+    /// trajectory.finish(state.clone()).unwrap();
+    /// assert_eq!(trajectory.number_of_action_points(), 2);
+    pub fn number_of_action_points(&self) -> usize{
+        self.states.len()
+    }
+    /// Register parameters for step. Provide starting state, action and payoff at start.
+    pub fn register_step_point(&mut self, state: S, agent: DP::AgentId, action: DP::ActionType, is_valid: bool)
+                               -> Result<(), AmfiteatrError<DP>>{
         if self.final_state.is_some(){
             return Err(TrajectoryError::UpdateOnFinishedGameTrajectory {
                 description: format!("{:?}", self)
@@ -276,6 +220,8 @@ impl<DP: DomainParameters, S: EnvironmentStateSequential<DP>> GameTrajectory<DP,
 
         Ok(())
     }
+    /// Registers state set and payoff in trajectory. Marks Trajectory as finished
+    /// and no later [`Self::register_step_point`] and [`Self::finish`] are allowed.
     pub fn finish(&mut self, state: S) -> Result<(), AmfiteatrError<DP>>{
         if self.final_state.is_some(){
             return Err(TrajectoryError::FinishingOnFinishedGameTrajectory {
@@ -284,13 +230,18 @@ impl<DP: DomainParameters, S: EnvironmentStateSequential<DP>> GameTrajectory<DP,
         self.final_state = Some(state);
         Ok(())
     }
+    /// Checks if trajectory is finished (the final game point is registered via method [`Self::finish`].
     pub fn is_finished(&self) -> bool{
         self.final_state.is_some()
     }
+    /// Checks if there is no step points registered.
     pub fn is_empty(&self) -> bool {
         self.states.is_empty() || self.validations.is_empty() || self.agents.is_empty() || self.actions.is_empty()
     }
-
+    /// Returns view for last step in trajectory - if the game is not finished it will be
+    /// step measured between two last registered entries. If game is finished (trajectory invoked
+    /// function [`GameTrajectory::finish`]) then the step between last registered information set
+    /// and information set provided for the finish function is used to mark last step.
     pub fn last_view_step(&self) -> Option<GameStepView<DP, S>>{
         if self.final_state.is_some(){
             self.view_step(self.states.len().saturating_sub(1))
@@ -298,6 +249,7 @@ impl<DP: DomainParameters, S: EnvironmentStateSequential<DP>> GameTrajectory<DP,
             self.view_step(self.states.len().saturating_sub(2))
         }
     }
+    /// Returns view of indexed step
     pub fn view_step(&self, index: usize) -> Option<GameStepView<DP, S>>{
         self.states.get(index+1).and_then(|se|{
             Some(GameStepView::new(