+++ amfiteatr_core/src/world/mod.rs
@@ -1,5 +1,0 @@
-mod generic;
-mod builder;
-
-pub use generic::*;
-pub use builder::*;